rename deeplink action
